refactor: custom hook refactor
