+++ src/hooks/useCategory.ts
@@ -1,8 +1,4 @@
-export const useCategory = (
-  currentCategory: any,
-  setCurrentCategory: any,
-  setOpenCategory: any,
-) => {
+const useCategory = (setCurrentCategory: any, setOpenCategory: any) => {
   const handleCategoryClick = (category: string) => {
     // 현재 카테고리와 다른 카테고리를 클릭했을 경우 카테고리 변경
     setCurrentCategory(category);
@@ -11,3 +7,5 @@ export const useCategory = (
 
   return { handleCategoryClick };
 };
+
+export default useCategory;
