+++ src/hooks/useLocationSearch.ts
@@ -1,0 +1,3 @@
+const useLocationSearch = () => {};
+
+export default useLocationSearch;
