+++ src/pages/Map/Map.tsx
@@ -1,4 +1,4 @@
-import { useEffect, useRef, useState } from 'react';
+import { useCallback, useEffect, useRef, useState } from 'react';
 import * as ReactDOMServer from 'react-dom/server';
 import { useParams } from 'react-router-dom';
 
@@ -9,7 +9,7 @@ import * as S from './Map.style';
 
 import InfoWrapper from './InfoWrapper/InfoWrapper';
 import Overlay from './Overlay/Overlay';
-import { useMap } from '../../hooks/useMap';
+import { useMap } from '../../hooks';
 
 const { kakao } = window;
 
@@ -43,6 +43,45 @@ export default function Map() {
     DB,
   );
 
+  // * 라우터 파라미터로 받은 bookstoreId 값에 따라 지도 이동
+  const moveMap = useCallback(
+    (bookstoreId: string | undefined) => {
+      // 지도가 생성되지 않았으면 함수 종료
+      if (!map) return;
+
+      // 이전에 클릭한 마커가 있으면 지도에서 제거
+      const prevOverlay = currentOverlayStoreId;
+      prevOverlay && prevOverlay.setMap(null);
+
+      // bookstoreId에 해당하는 마커로 지도 이동
+      DB.forEach((store) => {
+        if (store.ESNTL_ID === bookstoreId) {
+          const moveLatLon = new kakao.maps.LatLng(
+            store.FCLTY_LA,
+            store.FCLTY_LO,
+          );
+          map.setLevel(6); // 지도 확대 레벨 설정
+          map.panTo(moveLatLon); // 지도 중심 좌표 이동
+
+          // 커스텀 오버레이 생성
+          const overlayContent = ReactDOMServer.renderToString(
+            <Overlay info={store} />,
+          );
+          const overlay = new kakao.maps.CustomOverlay({
+            content: overlayContent,
+            map: map,
+            position: moveLatLon,
+            xAnchor: 0.5,
+            yAnchor: 0.5,
+          });
+
+          setCurrentOverlayStoreId(overlay);
+        }
+      });
+    },
+    [currentOverlayStoreId, DB, map],
+  );
+
   // * 첫 렌더링 시 지도 생성
   useEffect(() => {
     makeMap();
@@ -54,42 +93,10 @@ export default function Map() {
     makeMarkers();
   }, [DB, makeMarkers, markerImage]);
 
-  // * 라우터 파라미터로 받은 bookstoreId 값에 따라 지도 이동
+  // * bookstoreId가 변경되면 지도 이동
   useEffect(() => {
-    // 지도가 생성되지 않았으면 함수 종료
-    if (!map) return;
-
-    // 이전에 클릭한 마커가 있으면 지도에서 제거
-    const prevOverlay = currentOverlayStoreId;
-    prevOverlay && prevOverlay.setMap(null);
-
-    // bookstoreId에 해당하는 마커로 지도 이동
-    DB.forEach((store) => {
-      if (store.ESNTL_ID === bookstoreId) {
-        const moveLatLon = new kakao.maps.LatLng(
-          store.FCLTY_LA,
-          store.FCLTY_LO,
-        );
-        map.setLevel(6); // 지도 확대 레벨 설정
-        map.panTo(moveLatLon); // 지도 중심 좌표 이동
-
-        // 커스텀 오버레이 생성
-        const overlayContent = ReactDOMServer.renderToString(
-          <Overlay info={store} />,
-        );
-        const overlay = new kakao.maps.CustomOverlay({
-          content: overlayContent,
-          map: map,
-          position: moveLatLon,
-          xAnchor: 0.5,
-          yAnchor: 0.5,
-        });
-
-        setCurrentOverlayStoreId(overlay);
-      }
-    });
-    //eslint-disable-next-line
-  }, [bookstoreId, map, DB]);
+    moveMap(bookstoreId);
+  }, [moveMap, bookstoreId]);
 
   return (
     <S.Container>
